third_party/WebKit/LayoutTests/imported/wpt/html/semantics/selectors/pseudo-classes/inrange-outofrange-expected.txt

CONSOLE WARNING: line 59: The specified value "0" is not a valid email address.
This is a testharness.js-based test.
FAIL ':in-range' matches all elements that are candidates for constraint validation, have range limitations, and that are neither suffering from an underflow nor suffering from an overflow assert_array_equals: lengths differ, expected 10 got 16
PASS ':out-of-range' matches all elements that are candidates for constraint validation, have range limitations, and that are either suffering from an underflow or suffering from an overflow 
FAIL ':in-range' update number1's value < min assert_array_equals: lengths differ, expected 9 got 15
PASS ':out-of-range' update number1's value < min 
FAIL ':in-range' update number3's min < value assert_array_equals: lengths differ, expected 10 got 16
PASS ':out-of-range' update number3's min < value 
Harness: the test ran to completion.
